purgeData/requirements.txt
# Previous dependencies (keep all of these)
streamlit
langchain
langchain-community
langchain-ollama
pymysql
sqlalchemy
pandas
plotly
matplotlib
seaborn
python-dotenv
ollama
requests

# NEW: RAG learning system dependencies
scikit-learn>=1.3.0
numpy>=1.24.0

# Optional: For Excel file support (if you upload .xlsx files)
openpyxl>=3.1.0
xlrd>=2.0.1